requirements.txt
boto3
fastapi[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]==3.3.0
pydantic-settings
python-dotenv